requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
tenacity==8.2.3
python-multipart==0.0.6